requirements.txt
pandas
streamlit
altair
joblib
google-cloud-bigquery
db-dtypes
